requirements.txt
# Django stuff
Django==2.2.20
semver==2.6.0
pytz==2017.2
# PIL support
Pillow==8.1.1
python-dateutil==2.8.1

# reporting
pandas==1.1.5

# MySQL-python==1.2.5 # python3 incompatible
mysqlclient==2.0.1

django-storages==1.7

python-resize-image==1.1.19

# CSky
# django-cachemodel==2.1.7  # python3 incompatible
-e git://github.com/concentricsky/django-cachemodel.git@0f3081a2427c525f243119426432abd6484dc837#egg=django-cachemodel
# django-basic-models==3.0.4 # python3 incompatible
-e git://github.com/concentricsky/django-basic-models.git@50ae1448e61ee14941f502ec6d470a4799da001e#egg=django-basic-models


# DOCS
django-rest-swagger==2.2.0


django-object-actions==2.0.0

python-memcached==1.59

djangorestframework==3.11.2

# Django Allauth
django-allauth==0.37.1
## FW oauthlib==2.1.0
oauthlib==3.1.0
requests==2.24.0
## FW requests-oauthlib==0.4.2
requests-oauthlib==1.3.0
# Permissions framework
rules==0.4

# OAuth
django-oauth-toolkit==1.1.2

# AUTH
djangorestframework-simplejwt==4.7.1
password-strength==0.0.3.post2
django-graphql-jwt==0.3.2

# Some extras
django-cors-headers==1.1.0
django-cors-middleware==1.4.0
django-autoslug==1.9.6

# wsgiref==0.1.2 # python3 incompatible
sqlparse==0.2.2
netaddr
django-extensions==2.0.7

# Utilities for working with badges
jsonschema==2.6.0
simplejson==3.6.4

# JSON-LD
PyLD==0.7.1
rfc3987==1.3.4
pypng==0.0.18
jsonfield==2.0.2

# markdown support
Markdown==2.6.8
django-markdownify==0.1.0
bleach==3.3.0

# Open Badges related utilities
openbadges==1.1.0
openbadges-bakery==1.0.1

celery==4.1.1
django-celery-results==1.0.1
python-json-logger==0.1.2

# SSL Support
cffi==1.14.5
cryptography==3.3.2
enum34==1.1.6
idna==2.5
ipaddress==1.0.16
pyasn1==0.1.9
pycparser==2.14
## FW six==1.9.0
six==1.13.0

#svg 2 png support
cairocffi==0.9.0
CairoSVG==2.5.1
cssselect==0.9.2
tinycss==0.4
# pycairo==1.19.1

# Backport of Python 3.0 CSV package with Unicode support
backports.csv==1.0.4

# EncryptedCursorPagination dependencies
more-itertools==2.5.0

responses==0.8.1
## FW django_mock_queries==2.0.0
django_mock_queries==2.1.5

# lti consumer
lti==0.9.2

# Optional AWS dependencies
# boto==2.48.0
# boto3==1.5.6
# django-boto==0.3.12
# django-ses==0.8.4

# after python3 upgrade
social-auth-app-django==3.1.0

# graphql
graphene-django==2.8.2
aniso8601<=7p

# inline css
premailer==3.7.0